Performance optimization
Optimize the performance with 2 simple changes.

Added a canSleep variable to the drawmarker section stops it from running every tick unless it's needed.

Performance change 0.1-0.2ms.

Utilized the isInMarker to impose the same 500 ms delay registering if entered cirkle.

Performance change 0.1-0.2

When outside the draw distance the ms usage of the script is around 0.01 on a good computer prob 0.3 0.4 on a low end computer.