Major Update
Hopefully fixed the ability to Duplicate Vehicles. However if you restart the resource while a vehicle is out you can duplicated it once. I will keep the Pound Wait just in case for those that wish to use that as well.

Fixed being able to store different vehicles in other garage locations. Example: Being able to store an owned Aircraft & Car Garage.

Fixed being able to store a vehicle if not in Driver Seat.